Add rest of handlers with TODO items and move non-handler modules out of handlers directory
