+++ packages/extensions/experience-platform-connector/src/handleEvent.js
