--- packages/extensions/experience-platform-connector/src/handlers/config.js
@@ -1,4 +1,0 @@
-import { default as signInHandler } from './userSignIn';
-import { default as signOutHandler } from './userSignOut';
-
-export default [signInHandler, signOutHandler];
