--- packages/extensions/experience-platform-connector/src/handlers/index.js
@@ -1,1 +1,0 @@
-export { default as handleEvent } from './handleEvent';
