+++ packages/extensions/experience-platform-connector/src/handlers/shoppingCartView.js
@@ -1,0 +1,13 @@
+const canHandle = event => {
+  // TODO: Return true if this module's handler can handle the event
+  return;
+}
+
+const handle = (sdk, event) => {
+  // TODO: Handler logic
+}
+
+export default {
+  canHandle,
+  handle
+}
