+++ packages/extensions/experience-platform-connector/src/handlers/signIn.js
