+++ packages/extensions/experience-platform-connector/src/handlers/signOut.js
